functionality for price range filtering

+++ outfitted/models.py
@@ -37,10 +37,10 @@ class Product(models.Model):
     Images = models.CharField(max_length=255, unique=True, blank=True, null=True)
     name = models.CharField(max_length=255)
     description = models.TextField(max_length=255)
-    price = models.DecimalField(max_digits=6, decimal_places=2)
+    price = models.FloatField()
     ratings = models.FloatField()
     buys = models.IntegerField()
-    mrp = models.DecimalField(max_digits=6, decimal_places=2)
+    mrp = models.FloatField()
     discount = models.FloatField()
 
     title = models.TextField(max_length=255,  blank=True)
@@ -86,7 +86,7 @@ class Order(models.Model):
     )
     user = models.ForeignKey(User, on_delete=models.CASCADE)
     order_date = models.DateTimeField(auto_now_add=True)
-    total_amount = models.DecimalField(max_digits=6, decimal_places=2)
+    total_amount = models.FloatField()
     status = models.CharField(max_length=255)
 
 
@@ -94,7 +94,7 @@ class OrderItem(models.Model):
     id = models.AutoField(primary_key=True)
     product = models.ForeignKey(Product, on_delete=models.CASCADE)
     quantity = models.IntegerField()
-    price = models.DecimalField(max_digits=6, decimal_places=2)
+    price = models.FloatField()
     order = models.ForeignKey(Order, on_delete=models.CASCADE)
 
 
@@ -104,5 +104,5 @@ class Payment(models.Model):
     )
     order = models.ForeignKey(Order, on_delete=models.CASCADE)
     payment_date = models.DateTimeField(auto_now_add=True)
-    amount = models.DecimalField(max_digits=6, decimal_places=2)
+    amount = models.FloatField()
     method = models.CharField(max_length=255)
